# docs/api/external.openapi.yaml
openapi: 3.1.0
info:
  title: LawGround 사용자 API
  version: 0.1.0-draft
  description: 등록·단건 조회는 2주차 구현. 나머지는 검토용 초안이며 팀 합의/전체 구현 완료를 뜻하지 않는다. 모든 예시는 합성 데이터다.
servers:
- url: http://127.0.0.1:8080/api/v1
...
paths:
  /questions:
    post:
      operationId: createQuestion
      summary: createQuestion
      x-implementation-status: implemented-week-02
      security:
      - SessionCookie: []
      responses:
        '201':
          description: 생성 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
              example:
                subject: BROKER_LAW
                stem: '[개발용 합성 문제] 옳은 설명을 고르세요.'
                questionType: SELECT_CORRECT
                choices:
                - number: 1
                  text: 합성 설명 1
                - number: 2
                  text: 합성 설명 2
                - number: 3
                  text: 합성 설명 3
                - number: 4
                  text: 합성 설명 4
                - number: 5
                  text: 합성 설명 5
                providedAnswerNumber: 2
                referenceDate: '2025-10-25'
                id: 00000000-0000-4000-8000-000000000001
                visibility: PRIVATE
                answerSource: USER_PROVIDED
                revision: 1
                createdAt: '2026-10-03T13:00:00Z'
                updatedAt: '2026-10-03T13:00:00Z'
          headers:
            Location:
              schema:
                type: string
            ETag:
              schema: &id001
                type: string
                pattern: ^"[1-9][0-9]*"$
        '400':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VALIDATION_FAILED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '401':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: AUTHENTICATION_REQUIRED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '403':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ACCESS_DENIED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '409':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: DUPLICATE_QUESTION
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '413':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PAYLOAD_TOO_LARGE
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionInput'
            example:
              subject: BROKER_LAW
              stem: '[개발용 합성 문제] 옳은 설명을 고르세요.'
              questionType: SELECT_CORRECT
              choices:
              - number: 1
                text: 합성 설명 1
              - number: 2
                text: 합성 설명 2
              - number: 3
                text: 합성 설명 3
              - number: 4
                text: 합성 설명 4
              - number: 5
                text: 합성 설명 5
              providedAnswerNumber: 2
              referenceDate: '2025-10-25'
      parameters:
      - &id003
        name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 500
          pattern: \S
    get:
      operationId: listQuestions
      summary: listQuestions
      x-implementation-status: planned
      security:
      - SessionCookie: []
      responses:
        '200':
          description: 조회/처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
              example:
                items: []
                nextCursor: null
        '400':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VALIDATION_FAILED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '401':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: AUTHENTICATION_REQUIRED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '403':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ACCESS_DENIED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '404':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: RESOURCE_NOT_FOUND
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
      parameters:
      - &id006
        name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
      - &id007
        name: cursor
        in: query
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 1000
          pattern: \S
      - name: scope
        in: query
        required: false
        schema:
          type: string
          enum:
          - mine
          - public
          default: mine
  /questions/{id}:
    get:
      operationId: getQuestion
      summary: getQuestion
      x-implementation-status: implemented-week-02
      security:
      - {}
      - SessionCookie: []
      responses:
        '200':
          description: 조회/처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
              example:
                subject: BROKER_LAW
                stem: '[개발용 합성 문제] 옳은 설명을 고르세요.'
                questionType: SELECT_CORRECT
                choices:
                - number: 1
                  text: 합성 설명 1
                - number: 2
                  text: 합성 설명 2
                - number: 3
                  text: 합성 설명 3
                - number: 4
                  text: 합성 설명 4
                - number: 5
                  text: 합성 설명 5
                providedAnswerNumber: 2
                referenceDate: '2025-10-25'
                id: 00000000-0000-4000-8000-000000000001
                visibility: PRIVATE
                answerSource: USER_PROVIDED
                revision: 1
                createdAt: '2026-10-03T13:00:00Z'
                updatedAt: '2026-10-03T13:00:00Z'
          headers:
            ETag:
              schema: *id001
        '400':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VALIDATION_FAILED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '404':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: RESOURCE_NOT_FOUND
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
      parameters:
      - &id002
        name: id
        in: path
        required: true
        schema: &id005
          type: string
          format: uuid
    put:
      operationId: replaceQuestion
      summary: replaceQuestion
      x-implementation-status: planned
      security:
      - SessionCookie: []
      responses:
        '200':
          description: 조회/처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
              example:
                subject: BROKER_LAW
                stem: '[개발용 합성 문제] 옳은 설명을 고르세요.'
                questionType: SELECT_CORRECT
                choices:
                - number: 1
                  text: 합성 설명 1
                - number: 2
                  text: 합성 설명 2
                - number: 3
                  text: 합성 설명 3
                - number: 4
                  text: 합성 설명 4
                - number: 5
                  text: 합성 설명 5
                providedAnswerNumber: 2
                referenceDate: '2025-10-25'
                id: 00000000-0000-4000-8000-000000000001
                visibility: PRIVATE
                answerSource: USER_PROVIDED
                revision: 2
                createdAt: '2026-10-03T13:00:00Z'
                updatedAt: '2026-10-03T13:00:00Z'
          headers:
            ETag:
              schema: *id001
        '400':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VALIDATION_FAILED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '401':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: AUTHENTICATION_REQUIRED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '403':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ACCESS_DENIED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '404':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: RESOURCE_NOT_FOUND
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '409':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VERSION_CONFLICT
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '413':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PAYLOAD_TOO_LARGE
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '428':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PRECONDITION_REQUIRED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionInput'
            example:
              subject: BROKER_LAW
              stem: '[개발용 합성 문제] 옳은 설명을 고르세요.'
              questionType: SELECT_CORRECT
              choices:
              - number: 1
                text: 합성 설명 1
              - number: 2
                text: 합성 설명 2
              - number: 3
                text: 합성 설명 3
              - number: 4
                text: 합성 설명 4
              - number: 5
                text: 합성 설명 5
              providedAnswerNumber: 2
              referenceDate: '2025-10-25'
      parameters:
      - *id002
      - &id004
        name: If-Match
        in: header
        required: true
        schema: *id001
      - *id003
    delete:
      operationId: deleteQuestion
      summary: deleteQuestion
      x-implementation-status: planned
      security:
      - SessionCookie: []
      responses:
        '204':
          description: 처리 완료, 본문 없음
        '400':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VALIDATION_FAILED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '401':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: AUTHENTICATION_REQUIRED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '403':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ACCESS_DENIED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '404':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: RESOURCE_NOT_FOUND
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '409':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VERSION_CONFLICT
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
        '428':
          description: 요청을 처리하지 못함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PRECONDITION_REQUIRED
                message: 입력값을 확인해 주세요.
                requestId: 00000000-0000-4000-8000-000000000009
                timestamp: '2026-10-03T13:00:00Z'
                fieldErrors: []
      parameters:
      - *id002
      - *id004
      - *id003
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    Choice:
      type: object
      additionalProperties: false
      properties:
        number:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
          minLength: 1
          maxLength: 2000
          pattern: \S
      required:
      - number
      - text
    QuestionInput:
      type: object
      additionalProperties: false
      properties:
        subject: &id008
          type: string
          enum:
          - BROKER_LAW
        stem: &id009
          type: string
          minLength: 1
          maxLength: 10000
          pattern: \S
        questionType: &id010
          type: string
          enum:
          - SELECT_CORRECT
          - SELECT_INCORRECT
        choices: &id011
          type: array
          items:
            $ref: '#/components/schemas/Choice'
          minItems: 5
          maxItems: 5
          allOf:
          - contains:
              type: object
              properties:
                number:
                  const: 1
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 2
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 3
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 4
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 5
              required:
              - number
            minContains: 1
            maxContains: 1
        providedAnswerNumber: &id012
          type: integer
          minimum: 1
          maximum: 5
        referenceDate: &id013
          type: string
          format: date
      required:
      - subject
      - stem
      - questionType
      - choices
      - providedAnswerNumber
      - referenceDate
...
    FieldError:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 160
          pattern: \S
        message:
          type: string
          minLength: 1
          maxLength: 500
          pattern: \S
      required:
      - field
      - message
...
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
          - VALIDATION_FAILED
          - MALFORMED_REQUEST
          - REQUEST_FAILED
          - AUTHENTICATION_REQUIRED
          - ACCESS_DENIED
          - RESOURCE_NOT_FOUND
          - DUPLICATE_QUESTION
          - IDEMPOTENCY_CONFLICT
          - VERSION_CONFLICT
          - RESULT_NOT_READY
          - PAYLOAD_TOO_LARGE
          - PRECONDITION_REQUIRED
          - RATE_LIMITED
          - INTERNAL_SERVER_ERROR
        message:
          type: string
          minLength: 1
          maxLength: 500
          pattern: \S
        requestId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: \S
        timestamp: &id014
          type: string
          format: date-time
          pattern: Z$
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          minItems: 0
          maxItems: 50
      required:
      - code
      - message
      - requestId
      - timestamp
      - fieldErrors
    Question:
      type: object
      additionalProperties: false
      properties:
        subject:
          type: string
          enum:
          - BROKER_LAW
        stem:
          type: string
          minLength: 1
          maxLength: 10000
          pattern: \S
        questionType:
          type: string
          enum:
          - SELECT_CORRECT
          - SELECT_INCORRECT
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
          minItems: 5
          maxItems: 5
          allOf:
          - contains:
              type: object
              properties:
                number:
                  const: 1
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 2
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 3
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 4
              required:
              - number
            minContains: 1
            maxContains: 1
          - contains:
              type: object
              properties:
                number:
                  const: 5
              required:
              - number
            minContains: 1
            maxContains: 1
        providedAnswerNumber:
          type: integer
          minimum: 1
          maximum: 5
        referenceDate: &id015
          type: string
          format: date
        id: *id005
        visibility:
          type: string
          enum:
          - PRIVATE
          - PUBLIC
        answerSource:
          type: string
          enum:
          - USER_PROVIDED
          - OFFICIAL_VERIFIED
        revision: &id016
          type: integer
          minimum: 1
        createdAt: *id014
        updatedAt: *id014
      required:
      - subject
      - stem
      - questionType
      - choices
      - providedAnswerNumber
      - referenceDate
      - id
      - visibility
      - answerSource
      - revision
      - createdAt
      - updatedAt
    RunSummary:
      type: object
      additionalProperties: false
      properties:
        id: *id005
        status: &id017
          type: string
          enum:
          - QUEUED
          - RUNNING
          - COMPLETED
          - FAILED
          - CANCELLED
        verdict:
          anyOf:
          - &id018
            type: string
            enum:
            - SUPPORTED
            - HOLD
            - REJECTED
          - type: 'null'
        executionMode: &id019
          type: string
          enum:
          - MOCK
          - LIVE
      required:
      - id
      - status
      - verdict
      - executionMode
    QuestionListItem:
      type: object
      additionalProperties: false
      properties:
        id: *id005
        stemPreview:
          type: string
          minLength: 1
          maxLength: 120
          pattern: \S
        subject:
          type: string
          enum:
          - BROKER_LAW
        referenceDate: *id015
        visibility:
          type: string
          enum:
          - PRIVATE
          - PUBLIC
        revision: *id016
        createdAt: *id014
        latestOwnRun:
          anyOf:
          - $ref: '#/components/schemas/RunSummary'
          - type: 'null'
      required:
      - id
      - stemPreview
      - subject
      - referenceDate
      - visibility
      - revision
      - createdAt
      - latestOwnRun
...
    QuestionList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuestionListItem'
          minItems: 0
          maxItems: 50
        nextCursor:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 1000
            pattern: \S
          - type: 'null'
      required:
      - items
      - nextCursor
...